Attempt to fix TaplUninstallRemoveTest.testUninstallFromAllApps
Based on the video, it seems that existing app icons cause the long press menu
to be opened when the test bot is trying to scroll to open the AllApps page.
The fix is to clear the data so there are no icons that will interrupt said scrolling.

There might be 2 issues that are affecting this test based on my research. So if it still fails, that doesn't mean this change wasn't necessary. I think there might also be an issue with the test infrastructure.

Bug: 386231522
Test: Ran a go/abdt run to verify test isn't flaky. This does have 100% reproduction though.
Flag: EXEMPT test fix
Change-Id: Id2b1cf89eda39d1d921af0da544f92604c3f8591