Fix issue that recommendations were visible in landscape mode.
ag/30344959 showed the recommendations if available, to avoid the issue
of widgets list flashing when exiting search mode. However, we usually
hide recommendations in phone landscape due to the sticky header layout
not having enough space. In this fix, instead, we rebind recommendations
as before (to let the logic there decide whether to show or hide them),
and in order to fix the flashing problem that the earlier change was handling, we reset the recycler view.

Bug: 391231345
Flag: EXEMPT bugfix
Test: Manual, we don't have tests for interactions between screens
Change-Id: I535aa43cc553c9287c5d38ef7cde30f01e84d347