Fix bug with app pair 10:90 launch
Small fix for supporting the 10:90 split ratio. Previously, we detected the leftTop app by checking for left == 0 or top == 0; this led to a crash on app pair launch for 10:90 because, for example, the left pair's left bound was negative. Now we check for left <= 0 or top <= 0 instead.

Bug: 391866990
Flag: com.android.wm.shell.enable_flexible_two_app_split
Test: Visually correct when a 90:10 pair goes to recents and back
Change-Id: Ie22c12ea547b6d2dfeea4570ec32a79040cc45c2